perf(network)!: non-blocking Zakura block-sync reactor with split sequencer channels (#199)
* chore!: cherry pick flup changes to #164

* chore: tune parame

* ci: fix clippy too_many_arguments and ignore the structopt advisory

- allow(clippy::too_many_arguments) on the sequencer-control trace helper,
  matching the convention already used on the sibling constructors/tracers
- ignore RUSTSEC-2022-0104 (structopt maintenance mode) in deny.toml; it is a
  transitive dep via abscissa_core, alongside the structopt-chain advisories
  already ignored there

* test(zebrad): update stored config fixture for the 2 GiB block-byte budget

The flup+tune lowered DEFAULT_BS_MAX_INFLIGHT_BLOCK_BYTES from 8 GiB to 2 GiB,
which the acceptance `config_tests` (last_config_is_stored) compares against
the stored fixture. Regenerate the only drifted field, matching the in-place
update #164 made to the same fixture.

* test(network): full serving queue drops sends without disconnecting the peer

Guards the new non-blocking serving path (try_send_*): on a full per-peer
outbound queue, Block/BlocksDone/Status/RangeUnavailable are dropped with a
serve_queue_full metric and the peer is NOT cancelled — only a genuine send
error cancels. Forces Full via a two-slot queue (3 bodies + terminator > 2
slots), asserts the peer's cancellation token stays clear, and that serving
self-heals once the queue drains.

+++ zebra-network/src/zakura/testkit/mock_blocksync.rs
@@ -361,7 +361,7 @@ impl HarnessConfig {
     fn block_sync_config(&self) -> ZakuraBlockSyncConfig {
         ZakuraBlockSyncConfig {
             max_blocks_per_response: self.max_blocks_per_response,
-            max_inflight_requests: self.max_inflight,
+            max_inflight_requests: u32::from(self.max_inflight),
             max_inflight_block_bytes: u64::MAX,
             max_submitted_block_applies: usize::from(self.max_inflight)
                 .saturating_mul(
